feat: fixing findFoodByName endpoint and improving/adding some service rules

+++ src/daily-event/daily-event.service.ts
@@ -45,7 +45,12 @@ export class DailyEventService {
 	}
 
 	async findOne(id: number): Promise<DailyEventDto> {
-		return this.dailyEventRepository.findOne(id)
+        const dailyEvent = await this.dailyEventRepository.findOne(id)
+
+        if (!dailyEvent) throw new NotFoundException('Evento diário nao encontrado!')
+
+        return dailyEvent
+
 	}
 
 	async update(id: number, updateDailyEventDto: UpdateDailyEventDto) {
@@ -67,6 +72,9 @@ export class DailyEventService {
 	}
 
 	async remove(id: number): Promise<void> {
+		const dailyEvent = await DailyEvent.findByPk(id)
+        if (!dailyEvent) throw new NotFoundException('Evento diário não encontrado!')
+
 		return this.dailyEventRepository.remove(id)
 	}
